test(sampling): Improve test coverage
