+++ tests/js/spec/views/settings/projectSampling/index.spec.tsx
@@ -1,4 +1,4 @@
-import {screen, userEvent} from 'sentry-test/reactTestingLibrary';
+import {screen, userEvent, within} from 'sentry-test/reactTestingLibrary';
 
 import {SamplingRuleType} from 'sentry/types/sampling';
 import {SAMPLING_DOC_LINK} from 'sentry/views/settings/project/sampling/utils';
@@ -6,6 +6,8 @@ import {SAMPLING_DOC_LINK} from 'sentry/views/settings/project/sampling/utils';
 import {renderComponent} from './utils';
 
 describe('Sampling', function () {
+  const EMPTY_MESSAGE = 'There are no transaction rules to display';
+
   it('renders empty', function () {
     MockApiClient.addMockResponse({
       url: '/projects/org-slug/project-slug/',
@@ -13,18 +15,9 @@ describe('Sampling', function () {
       body: TestStubs.Project(),
     });
 
-    const {container} = renderComponent({withModal: false});
+    renderComponent({withModal: false});
 
-    // Title
-    expect(screen.getByText('Sampling')).toBeInTheDocument();
-
-    // SubTitle
-    expect(
-      screen.getByText(
-        /Here you can define what transactions count towards your quota without updating your SDK/
-      )
-    ).toBeInTheDocument();
-
+    // Subitle has the right link to the docs
     expect(
       screen.getByRole('link', {
         name: 'SDK sampling configuration',
@@ -32,7 +25,6 @@ describe('Sampling', function () {
     ).toHaveAttribute('href', SAMPLING_DOC_LINK);
 
     // Rules tabs
-    expect(screen.getByRole('tablist')).toHaveAttribute('aria-label', 'Rules tabs');
     expect(screen.getAllByRole('tab')).toHaveLength(2);
     expect(screen.getByRole('tab', {name: /Distributed Traces/})).toHaveAttribute(
       'aria-selected',
@@ -49,152 +41,145 @@ describe('Sampling', function () {
     expect(screen.getByText('Rate')).toBeInTheDocument();
 
     // Empty message is displayed
-    expect(
-      screen.getByText('There are no transaction rules to display')
-    ).toBeInTheDocument();
+    expect(screen.getByText(EMPTY_MESSAGE)).toBeInTheDocument();
 
     // Actions
-    expect(screen.getByRole('button', {name: 'Add Rule'})).toBeInTheDocument();
     expect(screen.getByRole('button', {name: 'Read Docs'})).toHaveAttribute(
       'href',
       SAMPLING_DOC_LINK
     );
-
-    expect(container).toSnapshot();
+    expect(screen.getByRole('button', {name: 'Add Rule'})).toBeInTheDocument();
   });
 
-  describe('renders with rules', function () {
-    beforeAll(() => {
-      MockApiClient.clearMockResponses();
-      MockApiClient.addMockResponse({
-        url: '/projects/org-slug/project-slug/',
-        method: 'GET',
-        body: TestStubs.Project({
-          dynamicSampling: {
-            rules: [
-              {
-                sampleRate: 0.2,
-                type: 'trace',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'glob',
-                      name: 'trace.release',
-                      value: ['1.2.3'],
-                    },
-                  ],
-                },
-                id: 40,
+  it('renders distributed traces tab', function () {
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'GET',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.2,
+              type: 'trace',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'glob',
+                    name: 'trace.release',
+                    value: ['1.2.3'],
+                  },
+                ],
               },
-              {
-                sampleRate: 0.2,
-                type: 'transaction',
-                condition: {
-                  op: 'and',
-                  inner: [
-                    {
-                      op: 'custom',
-                      name: 'event.legacy_browser',
-                      value: [
-                        'ie_pre_9',
-                        'ie9',
-                        'ie10',
-                        'ie11',
-                        'safari_pre_6',
-                        'opera_pre_15',
-                        'opera_mini_pre_8',
-                        'android_pre_4',
-                      ],
-                    },
-                  ],
-                },
-                id: 41,
-              },
-              {
-                sampleRate: 0.5,
-                type: 'transaction',
-                condition: {
-                  op: 'and',
-                  inner: [],
-                },
-                id: 42,
-              },
-            ],
-            next_id: 43,
-          },
-        }),
-      });
+              id: 40,
+            },
+          ],
+          next_id: 41,
+        },
+      }),
     });
 
-    it('Distributed traces tab', async function () {
-      const {container, router} = renderComponent({
-        withModal: false,
-        ruleType: SamplingRuleType.TRACE,
-      });
-
-      // Tab is active
-      expect(
-        await screen.findByRole('tab', {name: /Distributed Traces/})
-      ).toHaveAttribute('aria-selected', 'true');
-
-      // Tab description
-      expect(
-        screen.getByText(
-          'Using a Trace ID, select all Transactions distributed across multiple projects/services which match your conditions.'
-        )
-      ).toBeInTheDocument();
-
-      // Tab content
-      expect(screen.getByTestId('sampling-rule')).toHaveTextContent('If');
-
-      // Empty message is not displayed
-      expect(
-        screen.queryByText('There are no transaction rules to display')
-      ).not.toBeInTheDocument();
-
-      // Switch tabs
-      userEvent.click(screen.getByRole('tab', {name: /Individual Transactions/}));
-
-      // Transaction tab is pushed to the router
-      expect(router.push).toHaveBeenCalledWith(`${SamplingRuleType.TRANSACTION}/`);
-
-      expect(container).toSnapshot();
+    const {router} = renderComponent({
+      withModal: false,
+      ruleType: SamplingRuleType.TRACE,
     });
 
-    it('Individual Transactions tab', async function () {
-      const {container, router} = renderComponent({
-        withModal: false,
-        ruleType: SamplingRuleType.TRANSACTION,
-      });
+    const tracesTab = screen.getByRole('tab', {name: /Distributed Traces/});
+    // Tab is active
+    expect(tracesTab).toHaveAttribute('aria-selected', 'true');
+    // Tab has the correct number of rules badge
+    expect(within(tracesTab).getByText('1')).toBeInTheDocument();
 
-      // Tab is active
-      expect(
-        await screen.findByRole('tab', {name: /Individual Transactions/})
-      ).toHaveAttribute('aria-selected', 'true');
-
-      // Tab description
+    // Tab description
+    expect(
       screen.getByText(
-        'Select Transactions only within this project which match your conditions.'
-      );
+        'Using a Trace ID, select all Transactions distributed across multiple projects/services which match your conditions.'
+      )
+    ).toBeInTheDocument();
 
-      // Tab content
-      const rules = screen.getAllByTestId('sampling-rule');
-      expect(rules[0]).toHaveTextContent('If');
-      expect(rules[1]).toHaveTextContent('Else');
+    // Tab content
+    expect(screen.getAllByTestId('sampling-rule').length).toBe(1);
+    expect(screen.getByTestId('sampling-rule')).toHaveTextContent('If');
+    expect(screen.getByText('Release')).toBeInTheDocument();
+    expect(screen.getByText('1.2.3')).toBeInTheDocument();
+    expect(screen.getByText('20%')).toBeInTheDocument();
 
-      // Empty message is not displayed
-      expect(
-        screen.queryByText('There are no transaction rules to display')
-      ).not.toBeInTheDocument();
+    // Empty message is not displayed
+    expect(screen.queryByText(EMPTY_MESSAGE)).not.toBeInTheDocument();
 
-      // Switch tabs
-      userEvent.click(screen.getByRole('tab', {name: /Distributed Traces/}));
+    // Switch tabs
+    userEvent.click(screen.getByRole('tab', {name: /Individual Transactions/}));
 
-      // Distributed Traces tab is pushed to the router
-      expect(router.push).toHaveBeenCalledWith(`${SamplingRuleType.TRACE}/`);
+    // Transaction tab is pushed to the router
+    expect(router.push).toHaveBeenCalledWith(`${SamplingRuleType.TRANSACTION}/`);
+  });
 
-      expect(container).toSnapshot();
+  it('renders individual transactions tab', function () {
+    MockApiClient.addMockResponse({
+      url: '/projects/org-slug/project-slug/',
+      method: 'GET',
+      body: TestStubs.Project({
+        dynamicSampling: {
+          rules: [
+            {
+              sampleRate: 0.2,
+              type: 'transaction',
+              condition: {
+                op: 'and',
+                inner: [
+                  {
+                    op: 'eq',
+                    name: 'event.environment',
+                    value: ['prod'],
+                  },
+                ],
+              },
+              id: 41,
+            },
+            {
+              sampleRate: 0.5,
+              type: 'transaction',
+              condition: {
+                op: 'and',
+                inner: [],
+              },
+              id: 42,
+            },
+          ],
+          next_id: 43,
+        },
+      }),
     });
+
+    const {router} = renderComponent({
+      withModal: false,
+      ruleType: SamplingRuleType.TRANSACTION,
+    });
+
+    const transactionsTab = screen.getByRole('tab', {name: /Individual Transactions/});
+    // Tab is active
+    expect(transactionsTab).toHaveAttribute('aria-selected', 'true');
+    // Tab has the correct number of rules badge
+    expect(within(transactionsTab).getByText('2')).toBeInTheDocument();
+
+    // Tab description
+    screen.getByText(
+      'Select Transactions only within this project which match your conditions.'
+    );
+
+    // Tab content
+    const rules = screen.getAllByTestId('sampling-rule');
+    expect(rules.length).toBe(2);
+    expect(rules[0]).toHaveTextContent('IfEnvironmentprod20%');
+    expect(rules[1]).toHaveTextContent('Else50%');
+
+    // Empty message is not displayed
+    expect(screen.queryByText(EMPTY_MESSAGE)).not.toBeInTheDocument();
+
+    // Switch tabs
+    userEvent.click(screen.getByRole('tab', {name: /Distributed Traces/}));
+
+    // Distributed Traces tab is pushed to the router
+    expect(router.push).toHaveBeenCalledWith(`${SamplingRuleType.TRACE}/`);
   });
 });
