+++ tests/js/spec/views/settings/projectSampling/utils.tsx
@@ -4,34 +4,22 @@ import {initializeOrg} from 'sentry-test/initializeOrg';
 import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 
 import GlobalModal from 'sentry/components/globalModal';
+import {Organization} from 'sentry/types';
 import {SamplingRuleType} from 'sentry/types/sampling';
 import {OrganizationContext} from 'sentry/views/organizationContext';
 import Sampling from 'sentry/views/settings/project/sampling';
 
-export const commonConditionCategories = [
-  'Release',
-  'Environment',
-  'User Id',
-  'User Segment',
-  'Browser Extensions',
-  'Localhost',
-  'Legacy Browser',
-  'Web Crawlers',
-  'IP Address',
-  'Content Security Policy',
-  'Error Message',
-  'Transaction',
-];
-
 export function renderComponent({
+  orgOptions,
   ruleType = SamplingRuleType.TRACE,
   withModal = true,
 }: {
+  orgOptions?: Partial<Organization>;
   ruleType?: SamplingRuleType;
   withModal?: boolean;
 } = {}) {
   const {organization, project, router, routerContext} = initializeOrg({
-    organization: {
+    organization: orgOptions || {
       features: ['filters-and-sampling'],
     },
   } as Parameters<typeof initializeOrg>[0]);
@@ -59,13 +47,8 @@ export function renderComponent({
   return {container, router};
 }
 
-export async function renderModal(actionElement: HTMLElement, takeScreenshot = false) {
-  // Open Modal
+export async function openSamplingRuleModal(actionElement: HTMLElement) {
   userEvent.click(actionElement);
   const dialog = await screen.findByRole('dialog');
   expect(dialog).toBeInTheDocument();
-
-  if (takeScreenshot) {
-    expect(dialog).toSnapshot();
-  }
 }
